Separate string functions into separate headers
