+++ gemrb/core/CMakeLists.txt
@@ -105,7 +105,9 @@ FILE(GLOB gemrb_core_LIB_SRCS
 	System/Logger/Stdio.cpp
 	System/Logging.cpp
 	System/SlicedStream.cpp
+	System/CString.cpp
 	System/String.cpp
+	System/StringConversion.cpp
 	System/StringBuffer.cpp
 	System/swab.cpp
 	System/VFS.cpp
