+++ gemrb/plugins/GUIScript/GUIScript.cpp
@@ -68,9 +68,6 @@
 #include <algorithm>
 #include <cstdio>
 
-// MIPSPro fix for IRIX
-GEM_EXPORT size_t strlcpy(char *, const char *, size_t);
-
 using namespace GemRB;
 
 GUIScript *GemRB::gs = NULL;
